+++ django/applications/catmaid/static/libs/catmaid/error.js
@@ -121,4 +121,20 @@
    */
   CATMAID.NotImplementedError = class NotImplementedError extends CATMAID.Error {};
 
+  /**
+   * An error type to indicate an invalid user login.
+   */
+  CATMAID.InvalidLoginError = class InvalidLoginError extends CATMAID.Error {};
+
+  /**
+   * An error type to indicate an inactive user account. The optional meta field
+   * contains information on the reason of the inactivation.
+   */
+  CATMAID.InactiveLoginError = class InactiveLoginError extends CATMAID.Error {
+    constructor(message, detail, meta) {
+      super(message, detail);
+      this.meta = meta || {};
+    }
+  };
+
 })(CATMAID);