+++ django/projects/mysite/settings_base.py
@@ -279,7 +279,11 @@ CELERY_BEAT_SCHEDULE = {
     'daily-project-stats-summary-update': {
         'task': 'catmaid.tasks.update_project_statistics_from_scratch',
         'schedule': crontab(hour=23, minute=45)
-    }
+    },
+    'daily-inactivity-user-marking': {
+        'task': 'catmaid.tasks.deactivate_inactive_users',
+        'schedule': crontab(hour=00, minute=00)
+    },
 }
 
 # We use django-pipeline to compress and reference JavaScript and CSS files. To